+++ NEWS.md
@@ -1,5 +1,6 @@
-# kim 0.5.207
-* Added the following function(s): contingency_table, chi_squared_test
+# kim 0.5.212
+* Added the following function(s): contingency_table, chi_squared_test,
+geomean, mad_remove_outliers
 * Updated the following function(s) (and other functions that use them): 
 chi_squared_test_pairwise, odds_ratio